+++ Dtos/StudyCourseDtos/StudyCourseMobileStudentDetailResponseDto.cs
@@ -11,7 +11,7 @@ namespace griffined_api.Dtos.StudyCourseDtos
         public string Section { get; set; } = string.Empty;
         public string Course { get; set; } = string.Empty;
         public string? Level { get; set; }
-        public int TotalHour { get; set; }
+        public double TotalHour { get; set; }
         public string StartDate { get; set; } = string.Empty;
         public string EndDate { get; set; } = string.Empty;
         public Method Method { get; set; }
